requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

# Core Azure Functions runtime
azure-functions

# Azure SDKs
azure-identity
azure-storage-blob

# Database connectivity
pyodbc

# Data handling and transformation
pandas
numpy

# Machine Learning
scikit-learn
textblob

# Optional for NLP dependency of TextBlob
nltk

# Parquet support
pyarrow
fastparquet

# HTTP requests
requests
